Configure MassTransit with RabbitMQ and add consumer
New using directives for EventBus.Messages.Common, MassTransit, and Ordering.API.EventBusConsumer have been added to Program.cs. A scoped service for BasketOrderingConsumer has been added to the service collection. MassTransit has been configured to use RabbitMQ, including adding BasketOrderingConsumer as a consumer, configuring RabbitMQ with the host address specified in the configuration, and setting up a receive endpoint for the BasketCheckoutQueue to use BasketOrderingConsumer. The AddMassTransitHostedService method has been added to ensure that the MassTransit service is properly hosted.