Add nginx project with Docker and reverse proxy config
This commit introduces a new "nginx" project to the `Ecommerce.sln` solution. It includes the `nginx.Dockerfile` for building the Docker image and the `nginx.local.conf` file for configuring Nginx as a reverse proxy. The configuration routes requests to an IdentityServer instance and sets up necessary worker processes and event handling.