Add support for forwarded headers in pipeline
This commit introduces a new using directive for
`Microsoft.AspNetCore.HttpOverrides` and configures
`ForwardedHeadersOptions` in the application's request
pipeline. This allows the application to correctly
interpret `X-Forwarded-For` and `X-Forwarded-Proto`
headers, enhancing compatibility with reverse proxies
like Nginx. Known networks and proxies are cleared to
ensure only specified headers are processed.